refactor(islamic-website): move get news data func to utils
