+++ apps/islamic-website/app/page.tsx
@@ -1,7 +1,7 @@
 import { IndexPage } from '@haqq/islamic-website/index-page';
-import { mapStorybookToNews } from './news/page';
 import { storyblokInit, apiPlugin } from '@storyblok/js';
 import { Member } from '@haqq/islamic-ui-kit';
+import { getNewsPageContent } from '../utils/get-news-data';
 
 const STORYBLOK_ACCESS_TOKEN = process.env['STORYBLOK_ACCESS_TOKEN'];
 const VERCEL_ENV = process.env['VERCEL_ENV'];
@@ -24,27 +24,6 @@ function mapStoryblokToMembers(data: StoryblokMember[]): Member[] {
       url: member.url,
     };
   });
-}
-
-async function getNewsPageContent() {
-  const { storyblokApi } = storyblokInit({
-    accessToken: STORYBLOK_ACCESS_TOKEN,
-    use: [apiPlugin],
-  });
-
-  if (!storyblokApi) {
-    throw new Error('Failed to init storyblok');
-  }
-
-  const response = await storyblokApi.get('cdn/stories/media', {
-    version: VERCEL_ENV === 'production' ? 'published' : 'draft',
-  });
-
-  const posts = mapStorybookToNews(response.data.story.content.body[0].columns);
-
-  const mainPagePosts = posts.slice(0, 3);
-
-  return mainPagePosts;
 }
 
 async function getMembersContent() {
@@ -71,6 +50,7 @@ async function getMembersContent() {
 export default async function Page() {
   const news = await getNewsPageContent();
   const advisoryMembers = await getMembersContent();
+  
 
-  return <IndexPage news={news} advisoryMembers={advisoryMembers} />;
+  return <IndexPage news={news.slice(0,3)} advisoryMembers={advisoryMembers} />;
 }
